+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Unreleased: mitmproxy next
 * Addon to suppress unwanted error messages sent by mitmproxy. (@anneborcherding)
 * Updated imports and styles for web scanner helper addons. (@anneborcherding)
 * Inform when underscore-formatted options are used in client arg. (@jrblixt)
+* ASGIApp should ignore the HTTP flows loaded from somewhere. (@linw1995)
 * Binaries are now built with Python 3.9 (@mhils)
 * Fixed the web UI showing blank page on clicking details tab when server address is missing (@samhita-sopho)